+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [5.0.0] - TBD
 
+### Added
+
+- Add Statistical Data Analyzer Module ([#5622](https://github.com/wazuh/wazuh-qa/pull/5622)) \- (Framework + Tests)
+
 ## [4.10.0] - TBD
 
 ## [4.9.0] - TBD